+++ test/mocks/GlobalEffectAttack.sol
@@ -13,15 +13,15 @@ import {IMoveSet} from "../../src/moves/IMoveSet.sol";
 contract GlobalEffectAttack is IMoveSet {
     struct Args {
         Type TYPE;
-        uint256 STAMINA_COST;
-        uint256 PRIORITY;
+        uint32 STAMINA_COST;
+        uint32 PRIORITY;
     }
 
     IEngine immutable ENGINE;
     IEffect immutable EFFECT;
     Type immutable TYPE;
-    uint256 immutable STAMINA_COST;
-    uint256 immutable PRIORITY;
+    uint32 immutable STAMINA_COST;
+    uint32 immutable PRIORITY;
 
     constructor(IEngine _ENGINE, IEffect _EFFECT, Args memory args) {
         ENGINE = _ENGINE;
@@ -43,11 +43,11 @@ contract GlobalEffectAttack is IMoveSet {
         return false;
     }
 
-    function priority(bytes32) external view returns (uint256) {
+    function priority(bytes32) external view returns (uint32) {
         return PRIORITY;
     }
 
-    function stamina(bytes32) external view returns (uint256) {
+    function stamina(bytes32) external view returns (uint32) {
         return STAMINA_COST;
     }
