+++ test/mocks/SkipTurnMove.sol
@@ -12,14 +12,14 @@ import {IMoveSet} from "../../src/moves/IMoveSet.sol";
 contract SkipTurnMove is IMoveSet {
     struct Args {
         Type TYPE;
-        uint256 STAMINA_COST;
-        uint256 PRIORITY;
+        uint32 STAMINA_COST;
+        uint32 PRIORITY;
     }
 
     IEngine immutable ENGINE;
     Type immutable TYPE;
-    uint256 immutable STAMINA_COST;
-    uint256 immutable PRIORITY;
+    uint32 immutable STAMINA_COST;
+    uint32 immutable PRIORITY;
 
     constructor(IEngine _ENGINE, Args memory args) {
         ENGINE = _ENGINE;
@@ -39,11 +39,11 @@ contract SkipTurnMove is IMoveSet {
         return false;
     }
 
-    function priority(bytes32) external view returns (uint256) {
+    function priority(bytes32) external view returns (uint32) {
         return PRIORITY;
     }
 
-    function stamina(bytes32) external view returns (uint256) {
+    function stamina(bytes32) external view returns (uint32) {
         return STAMINA_COST;
     }
